requirements.txt
speedtest-cli
slackclient